Refactor Pitot class and it's usage
Change Pitot class to abstract where data is from. The class use to be linked with a ADC aquisition, what turned the implementation of aquisitions from the Arduino somehow bad. Now the class don't know where the data comes from, it just receives raw data, calculate some internal values and repass those values.

The ADC value is deprecated for now, since the data should come in the Voltage form. To abstract even more, data should come in the form of Pressure, because this is what the sensor is delivering in the end. The Main program should not need to know how to calculate Pressure from Voltage, since this change from sensor to sensor.

Pitot treatement on Main is now iterated, based on Configurador, not hard-coded. 

Apelido of pitots is shortened.