Change default value passed to Pitot from Voltage to Pressure
The pitot sensor, for default, measures differential pressure on it's main tube. Since the differencial pressure is in relation to a atmospheric pressure, the differential pressure read is a Dynamic Pressure (which is proportional to the fluid velocity).

The output of the pressure sensor is voltage, but this voltage is directly proportional (linear) to the differential pressure. As this conversion is fixed, the logical output from the Arduino that is reading the pitot sensor should be pressure, not voltage, and so the Raspberry should receive pressure, and not voltage. No conversion shoulb be made on the Raspberry side, it should abstract this information.

For this, the default variable on Pitot class is being changed from Voltage to Pressure.