getData() becomes updateData() and getData()
Previous getData() function pulled data from Arduino and returned to whoever's calling the getData() function. Now there's a updateData() function that pulls data from Arduino and put in a dictionary, but does not return it, and there's a getData() function that returns the dictionary of data.

This way the software can have a Thread updating the data (with updateData()), and the sensors itself pulling the data (with getData()).