Remove multiple port connection and insert port on object creation
The previous way consisted of trying to connect on several ports, but this will not work when both an Arduino *and* a Serial Transmitter are connected. To prevent future problems, now the Arduino class receives a port as a parameter, with a default one being used if not.

Also the entire atualiza() code  was put inside a try/excpet block, in case no Arduino is connected.