Implement SeletorDeModos()
SeletorDeModos is a new class with the aim to unite different modes of data transmission. It receives a setModo() command, change the value on criador.modo variable, and calls the setTransmissao()  method on each pertinent variable depending on the selected mode.

All the code on Ajudante class that was doing the same was removed and it is now implemented on SeletorDeModos().