Create SondaAoA sensor
This is a "pitot based" angle-of-attack probe.
This probe has two main calibration factors, based on the proper equation (see reference for aoa probe).
This sensors takes data from the defined pitots and calculate the AoA based on their dynamic/differential pressures.
The proper pitots and calibration factors can be configures on Configurador(), as usual.