Implement new usage routine
The new usage routine can be described in few steps:
1. User connect sensors and power cable;
2. Platform check for transmitter and start sending heartbeat signals;
3. Platform waits for settings from app;
4. User connects transmitter to cellphone, opens the app, verifies for the heartbeat signal, open connection and choose/send settings;
5. Platform receives settings and update it internally;
6. Platform go on with the initialization, but does not enable transmission and recording;
7. User enable transmission from the app;
8. User send calibration commands (pitot, cells, ultrasound, etc);
9. User enables recording;

Those steps were implemented mainly on telemetria.py and ajudante.py.

On telemetria.py the order of threads were changed, with the transmitter thread openning first, and the heartbeat and telecommand right after, this way the app is already working and can send commands. The recorder starts right after, with the settings updated. After that all the threads start, but recording and transmission wait for commands.